requirements.txt
# libraries
crewai
crewai_tools
cffi
cryptography

# environment variable and LLM management
python-dotenv
openai

# Observability of system
agentops




